LayoutTests/storage/websql/transaction-error-callback-isolated-world-expected.txt

CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
ALERT: undefined
ALERT: PASS: document.body.bar visible in a callback created in this world.
This test tests that the transaction error callback is called in the right world.
